Original change: https://android-review.googlesource.com/c/platform/packages/modules/Bluetooth/+/3073885

+++ system/stack/a2dp/a2dp_vendor_aptx.cc
@@ -362,7 +362,8 @@ std::string A2DP_VendorCodecInfoStringAptx(const uint8_t* p_codec_info) {
 
   a2dp_status = A2DP_ParseInfoAptx(&aptx_cie, p_codec_info, true);
   if (a2dp_status != A2DP_SUCCESS) {
-    res << "A2DP_ParseInfoAptx fail: " << loghex(a2dp_status);
+    res << "A2DP_ParseInfoAptx fail: "
+        << loghex(static_cast<uint8_t>(a2dp_status));
     return res.str();
   }
