+++ system/stack/a2dp/a2dp_vendor_aptx_hd.cc
@@ -383,7 +383,8 @@ std::string A2DP_VendorCodecInfoStringAptxHd(const uint8_t* p_codec_info) {
 
   a2dp_status = A2DP_ParseInfoAptxHd(&aptx_hd_cie, p_codec_info, true);
   if (a2dp_status != A2DP_SUCCESS) {
-    res << "A2DP_ParseInfoAptxHd fail: " << loghex(a2dp_status);
+    res << "A2DP_ParseInfoAptxHd fail: "
+        << loghex(static_cast<uint8_t>(a2dp_status));
     return res.str();
   }
